parallelize ENA uploads
