+++ conf/iridanext.config
@@ -17,7 +17,7 @@ iridanext {
         metadata {
             samples {
                 csv {
-                    path = "**/upload_metadata.csv"
+                    path = "**/*_upload_metadata.csv"
                     idcol = "library_name"
                 }
             }
