+++ modules/local/upload_reads_to_ena/main.nf
@@ -1,4 +1,5 @@
-process UPLOAD_TO_ENA {
+process UPLOAD_READS_TO_ENA {
+    tag "$meta.id"
     label 'process_single'
 
     conda "${moduleDir}/environment.yml"
@@ -7,7 +8,7 @@ process UPLOAD_TO_ENA {
         'biocontainers/ena-webin-cli:8.1.1--hdfd78af_1' }"
 
     input:
-    tuple path(multi_upload_manifest), path(reads)
+    tuple val(meta), path(reads), path(upload_manifest)
 
     output:
     path("ena_upload.log.txt")               , emit: upload_log
@@ -23,10 +24,10 @@ process UPLOAD_TO_ENA {
     # Temporary until upload_metadata.csv is generated by the upload_to_ena.py script
     touch upload_metadata.csv
 
-    upload_to_ena.py \\
+    upload_reads_to_ena.py \\
         --upload-username "${params.upload_username}" \\
         --upload-password "${params.upload_password}" \\
-        --multi-upload-manifest ${multi_upload_manifest} \\
+        --upload-manifest ${upload_manifest} \\
         --input-dir . \\
         ${test_flag} \\
         --upload-metadata upload_metadata.csv \\
@@ -34,7 +35,7 @@ process UPLOAD_TO_ENA {
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
-        upload_to_ena.py : 0.1.0
+        upload_reads_to_ena.py : \$(upload_reads_to_ena.py --version | awk '{print \$2}')
         ena-webin-cli : 8.1.1
     END_VERSIONS
     """
